+++ src/app/(service)/(my)/my-study/not-completed/page.tsx
@@ -5,7 +5,7 @@ import dynamic from 'next/dynamic';
 import { useState } from 'react';
 import Button from '@/components/common/ui/button';
 import Pagination from '@/components/common/ui/pagination';
-import NotCompletedGroupStudyList from '@/components/lists/not-completed-group-study-list';
+import NotCompletedGroupStudyList from '@/components/home/lists/not-completed-group-study-list';
 import { useAuthReady } from '@/features/auth/model/use-auth';
 import { useMemberStudyListV2Query } from '@/hooks/queries/user/use-member-study-list-query';
 import type { MemberStudyItem } from '@/types/api/group-study.types';
