+++ src/app/(service)/(my)/my-study/page.tsx
@@ -5,8 +5,8 @@ import dynamic from 'next/dynamic';
 import Link from 'next/link';
 import { useMemo } from 'react';
 import Button from '@/components/common/ui/button';
-import CompletedGroupStudyList from '@/components/lists/completed-group-study-list';
-import NotCompletedGroupStudyList from '@/components/lists/not-completed-group-study-list';
+import CompletedGroupStudyList from '@/components/home/lists/completed-group-study-list';
+import NotCompletedGroupStudyList from '@/components/home/lists/not-completed-group-study-list';
 import { useAuthReady } from '@/features/auth/model/use-auth';
 import { useMemberStudyListV2Query } from '@/hooks/queries/user/use-member-study-list-query';
 import type { MemberStudyItem } from '@/types/api/group-study.types';
