+++ src/app/(service)/(my)/notification/page.tsx
@@ -7,7 +7,7 @@ import type { GetMemberNotificationsTopicTypeEnum } from '@/api/openapi/api/noti
 import Button from '@/components/common/ui/button';
 import SingleDropdown from '@/components/common/ui/dropdown/single';
 import Pagination from '@/components/common/ui/pagination';
-import NotificationList from '@/components/lists/notification-list';
+import NotificationList from '@/components/home/lists/notification-list';
 import {
   useGetNotifications,
   useGetNotificationCategories,
