+++ src/components/home/card/study-card.tsx
