+++ src/components/home/contents/mission-detail-content.tsx
