+++ src/components/home/contents/payment-page-content.tsx
