+++ src/components/home/lists/completed-group-study-list.tsx
