+++ src/components/home/lists/group-study-list.tsx
