+++ src/components/home/lists/not-completed-group-study-list.tsx
