+++ src/components/home/lists/study-list-table.tsx
